api/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
joblib
scikit-learn
python-jose
python-multipart
pandas
requests